backend/requirements.txt
flask
flask-cors
python-dotenv
supabase
slack_sdk
openai
gunicorn